ArraySlice

ArraySlices are another way to use pointers. Instead of accessing addresses directly, they create another Array pointing to the address at the specified interval of the parent array.

Note

Read the PointerArithmetic page before using it

a = [1,2,3,4]
slice = ArraySlice(a, 2:3)

2-element Vector{Int64}:
 2
 3

Loading and Storing

Since it makes a new array, bounds checking are mandatory. Thus, accessing out of range elements will result in an error.

Performance

Because of the array creation, it doesn't need to access a structure to reach the pointer, but the memory already. It makes the ArraySlice to be faster than Slice and faster than a View, which is good for iterating over it.

Shift

Shifting works just like with Slices, and it can be a way out for the problem of accessing elements out of bounds.

Symbol	Description

<<	If positive, hides n elements at the beginning of the vector. Otherwise, it shows hidden elements there.

>>	If positive, increases the length of the array, showing new elements after the current last. Otherwise, it hides the n last shown elements.

PointerArithmetic.Pointers.shift_start — Method

shift_start(sl::AbstractArray, i::Integer)

Shifts the array starting point in a certain number of items. Can be called using << Usage:

julia> arr = [1,2,3,4]
julia> sl = ArraySlice(arr, 2:3)
2-element Vector{Int64}:
2
3
julia> sl << 1
1-element Vector{Int64}:
3
julia> sl << -1
3-element Vector{Int64}:
1
2
3

PointerArithmetic.ArraySlices.shift_end — Method

shift_end(sl::AbstractArray, i::Integer)

Shifts the array ending point in a certain number of items. Can be called using >> Usage:

julia> arr = [1,2,3,4]
julia> sl = ArraySlice(arr, 2:3)
2-element Vector{Int64}:
2
3
julia> sl >> 1
3-element Vector{Int64}:
2
3
4
julia> sl >> -1
1-element Vector{Int64}:
2