Pointer

Pointers are basically the same as using the Julia builtin function pointer(array). Instead of adding a new structure, it only adds C like accessing, storing and shifting operations.

Note

Read the first page before using it

a = [1,2,3,4]
ptr = Pointer(a)
ptr[3]

3

Also, you can shift the starting array item:

ptr = Pointer(a, 2)
ptr[1]

3

Loading and Storing

As mentioned earlier, storing and loading are similar to C syntax. Just like in C, you can access unbounded

a = [1,2,3,4]
ptr = Pointer(a, 2)
ptr[1]

3

ptr[-1]

1

Example of load operation

ptr[0] = 100
a

4-element Vector{Int64}:
   1
 100
   3
   4

Example of store operation

Note

Bounds are not problems here

Shift

Shifting a pointer is similar to memory address manipulation. Just like in C, you can now use the shift operator to add or subtract addresses from the base address in the pointer.

For this operation, there is one shifting operator:

Symbol	Description

<<	If positive, hides n elements at the beginning of the vector. Otherwise, it shows hidden elements there.

PointerArithmetic.ptr_shift_start — Method

ptr_shift_start(ptr::Ptr, i::Int)

Shifts the pointer in the specified i number of elements. Another way to call is to use ptr << nof_elements Usage:

julia> a = [1,2,3,4]
julia> pt = Pointer(a)
julia> pt[1]
1
julia> pt = pt << 2
julia> pt[1]
3