Slice

Slices are easiest way to use pointers. They are basically a data structure that holds all the needed data for pointer operations. By using this, you can create kind of a SubArray not limited by bounds. This means that something like this is perfectly possible:

a = [1,2,3,4]
slice = Slice(a, 2:3)

2-element Slice{Int64}:
 2
 3

As you can see, through the use of the pointer inside of slice it was able to access the memory directly by its address.

Note

Slices are the only subarray-like method that has bounds checking, which makes them safe.

Loading and Storing

Since Slices are sort of a SubArray, changing a value inside of it affects the parent array.

a = [1,2,3,4]
slice = Slice(a, 2:3)
slice[0]

1

Example of load operation

slice[0] = 100
a

4-element Vector{Int64}:
 100
   2
   3
   4

Example of store operation

Note

Remember, bounds are not problems here

Shift

Slices are part of an array, increasing and decreasing its size is the same as moving the pointer, and the length of the array.

For this operation, there are two shifting operators:

Symbol	Description

<<	If positive, hides n elements at the beginning of the vector. Otherwise, it shows hidden elements there.

>>	If positive, increases the length of the array, showing new elements after the current last. Otherwise, it hides the n last shown elements.

PointerArithmetic.Pointers.shift_start — Method

shift_start(sl::Slice, i::Integer)

Shifts the array starting point in a certain number of items. Can be called using << Usage:

julia> arr = [1,2,3,4]
julia> sl = Slice(arr, 2:3)
2-element Slice{Int64}:
2
3
julia> sl << 1
1-element Slice{Int64}:
3
julia> sl << -1
3-element Slice{Int64}:
1
2
3

PointerArithmetic.ArraySlices.shift_end — Method

shift_end(sl::Slice, i::Integer)

Shifts the array ending point in a certain number of items. Can be called using >> Usage:

julia> arr = [1,2,3,4]
julia> sl = Slice(arr, 2:3)
2-element Slice{Int64}:
2
3
julia> sl >> 1
3-element Slice{Int64}:
2
3
4
julia> sl >> -1
1-element Slice{Int64}:
2

Slices of slices

Another feature related to slices is the Slice of Slice feature. By calling Slice, passing the Slice instead of an Array, you can take fractions of the parent array, by using the slice relative indices.

a = [1,2,3,4]
b = Slice(a, 2:3)

2-element Slice{Int64}:
 2
 3

c = Slice(a, 1:3)

3-element Slice{Int64}:
 1
 2
 3

Note

It doesn't affect performance

Note

It is also not bounded to the slice, but to the parent array